+++ db/migrate/20180613191654_add_file_size_to_admin_exports.rb
@@ -1,0 +1,5 @@
+class AddFileSizeToAdminExports < ActiveRecord::Migration[5.2]
+  def change
+    add_column :admin_exports, :file_size, :bigint, null: false, default: 0
+  end
+end